Move monitor_redraw_frequency_seconds_values.py outside of ui folder
