+++ src/monitor_redraw_frequency_seconds_values.py
